Extract logic to open tasks.json buffer
