+++ crates/tasks_ui/Cargo.toml
@@ -24,6 +24,8 @@ picker.workspace = true
 project.workspace = true
 task.workspace = true
 serde.workspace = true
+tree-sitter.workspace = true
+tree-sitter-json.workspace = true
 ui.workspace = true
 util.workspace = true
 workspace.workspace = true
